+++ lib/domain/models/mode_buckets.dart
@@ -1,0 +1,11 @@
+import 'package:accredit/domain/models/source_item.dart';
+
+class ModeBuckets {
+  final List<SourceItem> playful;
+  final List<SourceItem> serious;
+
+  const ModeBuckets({
+    required this.playful,
+    required this.serious,
+  });
+}